prefer size_t

+++ src/packet.c
@@ -467,7 +467,7 @@ packet_x11_open(LIBSSH2_SESSION * session, unsigned char *data,
  */
 static inline int
 packet_authagent_open(LIBSSH2_SESSION * session, unsigned char *data,
-                      unsigned long datalen,
+                      size_t datalen,
                       packet_authagent_state_t *authagent_state)
 {
     int failure_code = SSH_OPEN_CONNECT_FAILED;
